fix: Move source-linked preset before variable resolution step
